Make sure the precision is int to keep 64bits happy
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@3590 48e7efb5-ca39-0410-a469-dd3cf9ba447f

+++ src/xbt/log.c
@@ -468,7 +468,7 @@ void xbt_log_init(int *argc,char **argv) {
 			
 		  if (strncmp(argv[i],"--log=",strlen("--log=")))
 		      WARN2("Option %.*s is deprecated and will disapear in the future. Use --log instead.",
-			    strchr(argv[i],'=')-argv[i],argv[i]);
+			    (int)(strchr(argv[i],'=')-argv[i]),argv[i]);
 
 		  opt=strchr(argv[i],'=');
 		  opt++;
